updated readme

+++ routes/api/product-routes.js
@@ -112,18 +112,18 @@ router.put('/:id', (req, res) => {
 router.delete('/:id', async (req, res) => {
   // delete one product by its `id` value
   try {
-    const productData = await Product.destroy({
+const checkCategory = await Product.findByPk(req.params.id);
+    if (!checkCategory) {
+      res.status(404).json({ message: 'No product found with this id!' });
+      return;
+    }
+    await Product.destroy({
       where: {
         id: req.params.id,
       },
     });
+    res.status(200).json({ message: `Product ${checkCategory.product_name} with id of ${req.params.id} deleted successfully!` });
 
-    if (!productData) {
-      res.status(404).json({ message: 'No product found with this id!' });
-      return;
-    }
-
-    res.status(200).json(productData);
   }
   catch (err) {
     res.status(500).json(err);
@@ -131,3 +131,5 @@ router.delete('/:id', async (req, res) => {
 });
 
 module.exports = router;
+
+
